@@ -27,48 +27,47 @@
     "@babel/preset-env": "^7.20.2",
     "@babel/preset-react": "^7.18.6",
     "@babel/runtime": "^7.20.6",
+    "@emotion/is-prop-valid": "^1.3.1",
     "@fortawesome/fontawesome-svg-core": "^6.2.1",
     "@fortawesome/free-regular-svg-icons": "^6.2.1",
     "@fortawesome/free-solid-svg-icons": "^6.2.1",
     "@fortawesome/react-fontawesome": "^0.2.0",
     "@loadable/babel-plugin": "^5.10.0",
     "@loadable/component": "^5.15.2",
-    "@webpack-cli/serve": "^1.7.0",
     "aria-query": "^5.1.3",
     "axios": "^1.7.2",
     "axios-extensions": "^3.1.6",
     "core-js": "^3.26.1",
     "css-loader": "^5.1.3",
+    "date-fns": "^4.1.0",
     "formik": "^2.2.9",
     "html-webpack-plugin": "^5.5.0",
+    "http-proxy-middleware": "^3.0.5",
     "jest-environment-jsdom": "^29.7.0",
     "leaflet": "^1.9.3",
     "leaflet-geosearch": "2.7.0",
     "lodash.clonedeep": "^4.5.0",
     "lodash.debounce": "^4.0.8",
-    "mobx": "^6.7.0",
-    "mobx-react": "^7.1.0",
+    "mobx": "^6.13.7",
+    "mobx-react": "^9.2.0",
     "moment": "^2.29.4",
     "moment-parseformat": "^4.0.0",
     "polished": "^4.2.2",
     "process": "^0.11.10",
     "prop-types": "^15.8.1",
-    "react": "^16.14.0",
-    "react-datepicker": "^2.14.1",
-    "react-dates": "^20.2.2",
-    "react-dom": "^16.14.0",
-    "react-interpolate-component": "^0.12.0",
-    "react-leaflet": "^2.8.0",
+    "react": "18.2",
+    "react-datepicker": "^8.3.0",
+    "react-dom": "18.2",
+    "react-leaflet": "^4.2.1",
     "react-markdown": "^8.0.4",
     "react-modal": "^3.16.1",
     "react-refresh": "^0.14.0",
     "react-router-dom": "^5.3.4",
     "react-select": "^5.10.1",
     "react-transition-group": "^2.9.0",
-    "react-with-direction": "^1.4.0",
     "remark-gfm": "^3.0.1",
     "style-loader": "^1.2.1",
-    "styled-components": "^5.3.6",
+    "styled-components": "^6.1.18",
     "terraformer-wkt-parser": "1.2.0",
     "uuid": "^9.0.0",
     "webpack": "^5.75.0",
