@@ -13,7 +13,7 @@
 import 'core-js/stable'
 
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 
 import App from './app'
 
@@ -26,4 +26,5 @@ if (BUILD === 'development') {
   // const axe = require('@axe-core/react')
   // window.setTimeout(() => axe(React, ReactDOM, 1500, {}), 1000)
 }
-ReactDOM.render(<App />, document.getElementById('root'))
+const root = createRoot(document.getElementById('root'))
+root.render(<App />)
