@@ -1,8 +1,9 @@
 import React from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
+import withCustomProps from '@/utils/withCustomProps'
 
-const SplashBg = styled.div`
+const SplashBg = withCustomProps(styled.div)`
   display: flex;
   position: fixed;
   top: 0;
