@@ -1,20 +1,21 @@
 {
-/**
- * This file is part of the Etsin service
- *
- * Copyright 2017-2018 Ministry of Education and Culture, Finland
- *
- *
- * @author    CSC - IT Center for Science Ltd., Espoo Finland <servicedesk@csc.fi>
- * @license   MIT
- */
+  /**
+   * This file is part of the Etsin service
+   *
+   * Copyright 2017-2018 Ministry of Education and Culture, Finland
+   *
+   *
+   * @author    CSC - IT Center for Science Ltd., Espoo Finland <servicedesk@csc.fi>
+   * @license   MIT
+   */
 }
 
 import React from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
+import withCustomProps from '@/utils/withCustomProps'
 
-const SplashBg = styled.div`
+const SplashBg = withCustomProps(styled.div)`
   display: flex;
   position: fixed;
   top: 0;
