@@ -45,7 +45,7 @@ const ModalFieldList = ({
             component={DeleteButton}
             type="button"
             onClick={() => remove(item.uiid)}
-            attribute={{ 'aria-label': 'qvain.general.buttons.remove' }}
+            attributes={{ 'aria-label': 'qvain.general.buttons.remove' }}
           />
         )}
       </ListItemButtonContainer>
