@@ -1,8 +1,9 @@
 import styled from 'styled-components'
 
 import checkColor from '../../../../styles/styledUtils'
+import withCustomProps from '@/utils/withCustomProps'
 
-const Label = styled.div`
+const Label = withCustomProps(styled.div)`
   padding: 0.3em 0.6em 0.4em;
   border-radius: 0.2em;
   background: ${props => (props.color ? checkColor(props.color) : props.theme.color.gray)};
