@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 
 export const StoresContext = React.createContext()
 
-export const StoresProvider = ({ store, children }) => (
+export const StoresProvider = ({ store, children = null }) => (
   <StoresContext.Provider value={store}>{children}</StoresContext.Provider>
 )
 
@@ -23,4 +23,5 @@ StoresProvider.defaultProps = {
 export const useStores = () => React.useContext(StoresContext)
 
 export const withStores = Component =>
+  // eslint-disable-next-line react/display-name
   React.forwardRef((props, ref) => <Component ref={ref} {...props} Stores={useStores()} />)
